Gfx patch: Fix consistency of Pepper's House
I'm building right now a precise map of Pepper's House. 
Almost all episode respect a similar design, exept this one:
The side of the house ( East ) had only one window. 
There should be three, and a bull-eye on the top. 
Also, the stairway in the main room is next to the lab, 
a door was missing, and also a part of the tree needs to
be visible near the entry door.
( Note: all lang will receive this update. )